casts for openbsd goodness

+++ tenders/common/elf.c
@@ -228,7 +228,7 @@ void elf_load(int bin_fd, const char *bin_name, uint8_t *mem, size_t mem_size,
      * e_entry must be non-zero and within range of our memory allocation.
      */
     if (ehdr->e_entry < p_min_loadaddr || ehdr->e_entry >= mem_size) {
-        warnx("%s: INV_EXE: e_entry does not fall within the memory allocation range (%lu < %lu || %lu >= %zu)", bin_name, ehdr->e_entry, p_min_loadaddr, ehdr->e_entry, mem_size);
+        warnx("%s: INV_EXE: e_entry does not fall within the memory allocation range (%llu < %llu || %llu >= %zu)", bin_name,  (unsigned long long)ehdr->e_entry, (unsigned long long)p_min_loadaddr, (unsigned long long)ehdr->e_entry, mem_size);
         goto mem_cleanup;
     }
     e_entry = ehdr->e_entry;
@@ -268,7 +268,7 @@ void elf_load(int bin_fd, const char *bin_name, uint8_t *mem, size_t mem_size,
             continue;
 
         if (p_vaddr < p_min_loadaddr) {
-            warnx("%s: INV_EXE: phdr[%u].p_vaddr outside of valid memory range (%lu < %lu)", bin_name, ph_i, p_vaddr, p_min_loadaddr);
+            warnx("%s: INV_EXE: phdr[%u].p_vaddr outside of valid memory range (%llu < %llu)", bin_name, ph_i, (unsigned long long)p_vaddr, (unsigned long long)p_min_loadaddr);
             goto mem_cleanup;
         }
         /*
@@ -277,7 +277,7 @@ void elf_load(int bin_fd, const char *bin_name, uint8_t *mem, size_t mem_size,
          * any surprises later.
          */
         if (p_vaddr < plast_vaddr) {
-            warnx("%s: INV_EXE: phdr[%u].p_vaddr is not in ascending order (%lu < %lu)", bin_name, ph_i, p_vaddr, plast_vaddr);
+            warnx("%s: INV_EXE: phdr[%u].p_vaddr is not in ascending order (%llu < %llu)", bin_name, ph_i, (unsigned long long)p_vaddr, (unsigned long long)plast_vaddr);
             goto mem_cleanup;
         }
         else
@@ -287,11 +287,11 @@ void elf_load(int bin_fd, const char *bin_name, uint8_t *mem, size_t mem_size,
          * and verify result is within range.
          */
         if (align_down(p_vaddr, p_align, &p_vaddr_start)) {
-            warnx("%s: INV_EXE: phdr[%u].p_vaddr is outside of memory range (align_down with %lu (p_vaddr) and %lu (p_align) failed)", bin_name, ph_i, p_vaddr, p_align);
+            warnx("%s: INV_EXE: phdr[%u].p_vaddr is outside of memory range (align_down with %llu (p_vaddr) and %llu (p_align) failed)", bin_name, ph_i, (unsigned long long)p_vaddr, (unsigned long long)p_align);
             goto mem_cleanup;
         }
         if (p_vaddr_start < p_min_loadaddr) {
-            warnx("%s: INV_EXE: phdr[%u].p_vaddr_start is outside of memory range (%lu < %lu (p_min_loadaddr))", bin_name, ph_i, p_vaddr_start, p_min_loadaddr);
+            warnx("%s: INV_EXE: phdr[%u].p_vaddr_start is outside of memory range (%llu < %llu (p_min_loadaddr))", bin_name, ph_i, (unsigned long long)p_vaddr_start, (unsigned long long)p_min_loadaddr);
             goto mem_cleanup;
         }
         /*
@@ -299,22 +299,22 @@ void elf_load(int bin_fd, const char *bin_name, uint8_t *mem, size_t mem_size,
          * the Solo5 toolchains do not produce such executables.
          */
         if (p_vaddr_start < e_end) {
-            warnx("%s: INV_EXE: phdr[%u].p_vaddr_start is outside of memory range (%lu < %lu (e_end))", bin_name, ph_i, p_vaddr_start, e_end);
+            warnx("%s: INV_EXE: phdr[%u].p_vaddr_start is outside of memory range (%llu < %llu (e_end))", bin_name, ph_i, (unsigned long long)p_vaddr_start, (unsigned long long)e_end);
             goto mem_cleanup;
         }
         /*
          * Verify p_vaddr + p_filesz is within range.
          */
         if (p_vaddr >= mem_size) {
-            warnx("%s: INV_EXE: phdr[%u] program file segment falls outside of valid range (%lu >= %zu (mem_size))", bin_name, ph_i, p_vaddr, mem_size);
+            warnx("%s: INV_EXE: phdr[%u] program file segment falls outside of valid range (%llu >= %zu (mem_size))", bin_name, ph_i, (unsigned long long)p_vaddr, mem_size);
             goto mem_cleanup;
         }
         if (add_overflow(p_vaddr, p_filesz, temp)) {
-            warnx("%s: INV_EXE: phdr[%u] program file segment falls outside of valid range (%lu (p_vaddr) + %lu (p_filesz) overflowed)", bin_name, ph_i, p_vaddr, p_filesz);
+            warnx("%s: INV_EXE: phdr[%u] program file segment falls outside of valid range (%llu (p_vaddr) + %llu (p_filesz) overflowed)", bin_name, ph_i, (unsigned long long)p_vaddr, (unsigned long long)p_filesz);
             goto mem_cleanup;
         }
         if (temp > mem_size) {
-            warnx("%s: INV_EXE: phdr[%u] program file segment falls outside of valid range (%lu (p_vaddr + p_filesz) > %zu (mem_size))", bin_name, ph_i, temp, mem_size);
+            warnx("%s: INV_EXE: phdr[%u] program file segment falls outside of valid range (%llu (p_vaddr + p_filesz) > %zu (mem_size))", bin_name, ph_i, (unsigned long long)temp, mem_size);
             goto mem_cleanup;
         }
         /*
@@ -322,19 +322,19 @@ void elf_load(int bin_fd, const char *bin_name, uint8_t *mem, size_t mem_size,
          * alignment and verify result is within range.
          */
         if (p_memsz < p_filesz) {
-            warnx("%s: INV_EXE: phdr[%u] program memory segment falls outside of valid range (%lu (p_memsz) < %lu (p_filesz))", bin_name, ph_i, p_memsz, p_filesz);
+            warnx("%s: INV_EXE: phdr[%u] program memory segment falls outside of valid range (%llu (p_memsz) < %llu (p_filesz))", bin_name, ph_i, (unsigned long long)p_memsz, (unsigned long long)p_filesz);
             goto mem_cleanup;
         }
         if (add_overflow(p_vaddr, p_memsz, p_vaddr_end)) {
-            warnx("%s: INV_EXE: phdr[%u] program memory segment falls outside of valid range (%lu (p_vaddr) + %lu (p_memsz) overflowed)", bin_name, ph_i, p_vaddr, p_memsz);
+            warnx("%s: INV_EXE: phdr[%u] program memory segment falls outside of valid range (%llu (p_vaddr) + %llu (p_memsz) overflowed)", bin_name, ph_i, (unsigned long long)p_vaddr, (unsigned long long)p_memsz);
             goto mem_cleanup;
         }
         if (align_up(p_vaddr_end, p_align, &p_vaddr_end)) {
-            warnx("%s: INV_EXE: phdr[%u] program memory segment falls outside of valid range (align_up with %lu (p_vaddr_end) and %lu (p_align) failed)", bin_name, ph_i, p_vaddr_end, p_align);
+            warnx("%s: INV_EXE: phdr[%u] program memory segment falls outside of valid range (align_up with %llu (p_vaddr_end) and %llu (p_align) failed)", bin_name, ph_i, (unsigned long long)p_vaddr_end, (unsigned long long)p_align);
             goto mem_cleanup;
         }
         if (p_vaddr_end > mem_size) {
-            warnx("%s: INV_EXE: phdr[%u] program memory segment falls outside of valid range (%lu (p_vaddr + p_memsz) > %zu (mem_size))", bin_name, ph_i, p_vaddr_end, mem_size);
+            warnx("%s: INV_EXE: phdr[%u] program memory segment falls outside of valid range (%llu (p_vaddr + p_memsz) > %zu (mem_size))", bin_name, ph_i, (unsigned long long)p_vaddr_end, mem_size);
             goto mem_cleanup;
         }
         /*
@@ -361,7 +361,7 @@ void elf_load(int bin_fd, const char *bin_name, uint8_t *mem, size_t mem_size,
          * Double check result for host (caller) address space overflow.
          */
         if (host_vaddr < (mem + p_min_loadaddr)) {
-          warnx("%s: INV_EXE: phdr[%u] address space overflow (host_vaddr < (mem + p_min_loaddr))", bin_name, ph_i);
+            warnx("%s: INV_EXE: phdr[%u] address space overflow (host_vaddr < (mem + p_min_loaddr))", bin_name, ph_i);
             goto mem_cleanup;
         }
         nbytes = pread_in_full(bin_fd, host_vaddr, p_filesz,
@@ -371,7 +371,7 @@ void elf_load(int bin_fd, const char *bin_name, uint8_t *mem, size_t mem_size,
             goto mem_cleanup;
         }
         if (nbytes != p_filesz) {
-            warnx("%s: phdr[%u] host file segment mismatched (pread_in_full returned %zu != %lu (p_filesz))", bin_name, ph_i, nbytes, p_filesz);
+            warnx("%s: phdr[%u] host file segment mismatched (pread_in_full returned %zu != %llu (p_filesz))", bin_name, ph_i, nbytes, (unsigned long long)p_filesz);
             goto mem_cleanup;
         }
         memset(host_vaddr + p_filesz, 0, p_memsz - p_filesz);
@@ -383,11 +383,11 @@ void elf_load(int bin_fd, const char *bin_name, uint8_t *mem, size_t mem_size,
          * size.
          */
         if (p_vaddr_start & (EM_PAGE_SIZE - 1)) {
-            warnx("%s: INV_EXE: phdr[%u] p_vaddr_start not aligned with architectural page size (%lu & (%u - 1))", bin_name, ph_i, p_vaddr_start, EM_PAGE_SIZE - 1);
+            warnx("%s: INV_EXE: phdr[%u] p_vaddr_start not aligned with architectural page size (%llu & (%u - 1))", bin_name, ph_i, (unsigned long long)p_vaddr_start, EM_PAGE_SIZE - 1);
             goto mem_cleanup;
         }
         if (p_vaddr_end & (EM_PAGE_SIZE - 1)) {
-            warnx("%s: INV_EXE: phdr[%u] p_vaddr_end not aligned with architectural page size (%lu & (%u - 1))", bin_name, ph_i, p_vaddr_end, EM_PAGE_SIZE - 1);
+            warnx("%s: INV_EXE: phdr[%u] p_vaddr_end not aligned with architectural page size (%llu & (%u - 1))", bin_name, ph_i, (unsigned long long)p_vaddr_end, EM_PAGE_SIZE - 1);
             goto mem_cleanup;
         }
         int prot = PROT_NONE;
@@ -482,7 +482,7 @@ int elf_load_note(int bin_fd, const char *bin_name, uint32_t note_type,
              * p_filesz is less than minimum possible size of a NOTE header,
              * reject the executable.
              */
-            warnx("%s: INV_EXE: phdr[%u].p_filesz note does not fall within valid size (%lu < %lu)", bin_name, ph_i, phdr[ph_i].p_filesz, sizeof(Elf64_Nhdr));
+            warnx("%s: INV_EXE: phdr[%u].p_filesz note does not fall within valid size (%llu < %lu)", bin_name, ph_i, (unsigned long long)phdr[ph_i].p_filesz, sizeof(Elf64_Nhdr));
             goto mem_cleanup;
         }
         if (phdr[ph_i].p_filesz < sizeof nhdr)
@@ -525,7 +525,7 @@ int elf_load_note(int bin_fd, const char *bin_name, uint32_t note_type,
             goto mem_cleanup;
         }
         if (phdr[ph_i].p_filesz < sizeof nhdr + nhdr.h.n_descsz) {
-            warnx("%s: INV_EXE: phdr[%u] note does not fall within valid size (%lu < %lu)", bin_name, ph_i, phdr[ph_i].p_filesz, sizeof nhdr + nhdr.h.n_descsz);
+            warnx("%s: INV_EXE: phdr[%u] note does not fall within valid size (%llu < %lu)", bin_name, ph_i, (unsigned long long)phdr[ph_i].p_filesz, sizeof nhdr + nhdr.h.n_descsz);
             goto mem_cleanup;
         }
